fix(icon): move icon-component to data-folder
