+++ src/components/data/Icon/interfaces/Icon.interface.ts
