+++ src/components/data/Icon/stories/props.ts
