+++ src/components/data/Icon/styles.ts
