+++ src/components/data/Icon/types/Icon.Appearance.type.ts
