+++ src/components/data/Icon/types/Icon.Variant.type.ts
